+++ apps/web/src/actions/categories/DeleteCategory.ts
