+++ apps/web/src/actions/etl/categories/DeleteAllLookupCategories.ts
