+++ apps/web/src/actions/etl/categories/DeleteLookupCategory.ts
