+++ apps/web/src/components/transactions/EditProcessedTransaction.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import type { ProcessedTransaction } from "@/actions/transactions/parseTransaction";
-import { SearchableSelect } from "@coin-guard/ui";
+import type { ProcessedTransaction } from "@/actions/transactions/ParseTransaction";
 import { useGetCategories } from "@/hooks/categories/useGetCategories";
 import { processedTransactionsAtom } from "@/store/transactionsStore";
+import { SearchableSelect } from "@coin-guard/ui";
 import { useSetAtom } from "jotai";
 import { useCallback, useMemo } from "react";
