+++ apps/web/src/store/transactionsStore.ts
@@ -1,4 +1,4 @@
-import type { ProcessedTransaction } from "@/actions/transactions/parseTransaction";
+import type { ProcessedTransaction } from "@/actions/transactions/ParseTransaction";
 import { atom } from "jotai";
 
 export const processedTransactionsAtom = atom<ProcessedTransaction[]>([]);
